requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.7
python-dotenv>=1.0
rapidfuzz>=3.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
openai>=1.50
python-multipart>=0.0.6
